Stop fetching the battery level for every animation frame.
We need it only once per showing.  Also adjust the percentage
'calculation' a bit -- it's very dumb at the moment.

Since my observations have seen the raw value be around the 630-640 mark
when full and maybe around 530 when empty (need to confirm) I'm just
subtracting 540 from that value to get a percentage.  But ideally the
full range, and it's curve assuming it's not a linear discharge, would
be used to come up with a percentage full value.