Improve the reading of the time from serial with the lines face.
The countdown timer and a 'read time' message is drawn to the screen
before it attempts to read the time from the serial port.

Now, if you unplug the watch before then it won't attempt to set the
time, and keep the time it already had.

But to print the messages to the screen I had to pull the line, number
and letter drawing routines from the lines face code into it's own
module to make it available in general.