Remove the periodic battery check.
Just do it every time we refresh the watch face.

I originally moved the check out of the logic for showing the time since
when it was animated it was grabbing the batt% more often than
necessary.  But then I introduced a bug where I forgot that millis()
doesn't increment while sleeping so my checks to refresh the batt% were
never triggering.

We could keep the periodic check there but reduce the refresh period to
something quite low while keeping in mind that period is only counting
while it's awake.  But for now I'm just grabbing the value every time we
refresh the time since it's not currently animating anyway.